Minor changes from review feedback
